Removes unused nav code
