+++ package.json
@@ -40,7 +40,6 @@
     "react-dom": "^16.4.1",
     "react-md-spinner": "^0.3.0",
     "react-redux": "^5.0.7",
-    "react-router-dom": "^4.3.1",
     "react-scripts": "^2.1.1",
     "redux": "^4.0.0",
     "redux-thunk": "^2.3.0",
