--- src/components/NestedSideNav/index.jsx
@@ -1,3 +1,0 @@
-import NestedSideNav from './NestedSideNav'
-
-export default NestedSideNav
