--- src/components/SidebarNav/SideNav.css
@@ -1,20 +1,0 @@
-aside.sidebar {
-  width: 54px;
-}
-
-aside.sidebar nav > ul > li button.main img, aside.sidebar nav > ul > li button.main {
-  width: 32px;
-  height: 32px;
-}
-
-aside.sidebar nav > ul > li a.active button.main, aside.sidebar nav > ul > li a.active button.main:focus {
-  /*border-left: 3px solid #4A90E2;*/
-  opacity: 1;
-}
-
-aside.sidebar nav .submenu-container {
-  left: 60px;
-}
-aside.sidebar nav .submenu-container section header{
-  padding-bottom: 30px;
-}
