--- src/components/SidebarNav/index.js
@@ -1,3 +1,0 @@
-import SideNav from './SideNav'
-
-export default SideNav
